changed the logic of adding and sellling quantity

+++ lib/models/product_model.dart
@@ -8,7 +8,7 @@ class ProductModel {
   final int sellingPrice;
   String? description;
   dynamic productImage;
-  double productQuantity;
+  int productQuantity;
   ProductModel(
       {required this.category,
       required this.productName,
